+++ runner/monitor/CHANGELOG.md
@@ -11,6 +11,7 @@
 **Breaking Changes**
 
 **API Changes**
+* Move androidx.test.platform.tracing back to an internal API
 
 **Breaking API Changes**